# langya/platform/app/langya_platform_app.pb.gw.go
// Code generated by protoc-gen-grpc-gateway. DO NOT EDIT.
// source: langya_platform_app.proto

/*
Package app is a reverse proxy.

It translates gRPC into RESTful JSON APIs.
*/
package app

import (
	"context"
	"io"
	"net/http"

	"github.com/grpc-ecosystem/grpc-gateway/v2/runtime"
	"github.com/grpc-ecosystem/grpc-gateway/v2/utilities"
	"google.golang.org/grpc"
	"google.golang.org/grpc/codes"
	"google.golang.org/grpc/grpclog"
	"google.golang.org/grpc/metadata"
	"google.golang.org/grpc/status"
	"google.golang.org/protobuf/proto"
)

// Suppress "imported and not used" errors
var _ codes.Code
var _ io.Reader
var _ status.Status
var _ = runtime.String
var _ = utilities.NewDoubleArray
var _ = metadata.Join

func request_LangYaPlatform_ServiceContractBook_0(ctx context.Context, marshaler runtime.Marshaler, client LangYaPlatformClient, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
	var protoReq ContractBookReq
	var metadata runtime.ServerMetadata

	msg, err := client.ServiceContractBook(ctx, &protoReq, grpc.Header(&metadata.HeaderMD), grpc.Trailer(&metadata.TrailerMD))
	return msg, metadata, err

}

func local_request_LangYaPlatform_ServiceContractBook_0(ctx context.Context, marshaler runtime.Marshaler, server LangYaPlatformServer, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
	var protoReq ContractBookReq
	var metadata runtime.ServerMetadata

	msg, err := server.ServiceContractBook(ctx, &protoReq)
	return msg, metadata, err

}

func request_LangYaPlatform_ServiceContractBookSave_0(ctx context.Context, marshaler runtime.Marshaler, client LangYaPlatformClient, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
	var protoReq ContractBookReq
	var metadata runtime.ServerMetadata

	newReader, berr := utilities.IOReaderFactory(req.Body)
	if berr != nil {
		return nil, metadata, status.Errorf(codes.InvalidArgument, "%v", berr)
	}
	if err := marshaler.NewDecoder(newReader()).Decode(&protoReq); err != nil && err != io.EOF {
		return nil, metadata, status.Errorf(codes.InvalidArgument, "%v", err)
	}

	msg, err := client.ServiceContractBookSave(ctx, &protoReq, grpc.Header(&metadata.HeaderMD), grpc.Trailer(&metadata.TrailerMD))
	return msg, metadata, err

}

func local_request_LangYaPlatform_ServiceContractBookSave_0(ctx context.Context, marshaler runtime.Marshaler, server LangYaPlatformServer, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
	var protoReq ContractBookReq
	var metadata runtime.ServerMetadata

	newReader, berr := utilities.IOReaderFactory(req.Body)
	if berr != nil {
		return nil, metadata, status.Errorf(codes.InvalidArgument, "%v", berr)
	}
	if err := marshaler.NewDecoder(newReader()).Decode(&protoReq); err != nil && err != io.EOF {
		return nil, metadata, status.Errorf(codes.InvalidArgument, "%v", err)
	}

	msg, err := server.ServiceContractBookSave(ctx, &protoReq)
	return msg, metadata, err

}

// RegisterLangYaPlatformHandlerServer registers the http handlers for service LangYaPlatform to "mux".
// UnaryRPC     :call LangYaPlatformServer directly.
// StreamingRPC :currently unsupported pending https://github.com/grpc/grpc-go/issues/906.
// Note that using this registration option will cause many gRPC library features to stop working. Consider using RegisterLangYaPlatformHandlerFromEndpoint instead.
func RegisterLangYaPlatformHandlerServer(ctx context.Context, mux *runtime.ServeMux, server LangYaPlatformServer) error {

	mux.Handle("GET", pattern_LangYaPlatform_ServiceContractBook_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
		ctx, cancel := context.WithCancel(req.Context())
		defer cancel()
		var stream runtime.ServerTransportStream
		ctx = grpc.NewContextWithServerTransportStream(ctx, &stream)
		inboundMarshaler, outboundMarshaler := runtime.MarshalerForRequest(mux, req)
		var err error
		ctx, err = runtime.AnnotateIncomingContext(ctx, mux, req, "/app.LangYaPlatform/ServiceContractBook", runtime.WithHTTPPathPattern("/contract/book"))
		if err != nil {
			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
			return
		}
		resp, md, err := local_request_LangYaPlatform_ServiceContractBook_0(ctx, inboundMarshaler, server, req, pathParams)
		md.HeaderMD, md.TrailerMD = metadata.Join(md.HeaderMD, stream.Header()), metadata.Join(md.TrailerMD, stream.Trailer())
		ctx = runtime.NewServerMetadataContext(ctx, md)
		if err != nil {
			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
			return
		}

		forward_LangYaPlatform_ServiceContractBook_0(ctx, mux, outboundMarshaler, w, req, resp, mux.GetForwardResponseOptions()...)

	})

	mux.Handle("POST", pattern_LangYaPlatform_ServiceContractBookSave_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
		ctx, cancel := context.WithCancel(req.Context())
		defer cancel()
		var stream runtime.ServerTransportStream
		ctx = grpc.NewContextWithServerTransportStream(ctx, &stream)
		inboundMarshaler, outboundMarshaler := runtime.MarshalerForRequest(mux, req)
		var err error
		ctx, err = runtime.AnnotateIncomingContext(ctx, mux, req, "/app.LangYaPlatform/ServiceContractBookSave", runtime.WithHTTPPathPattern("/contract/book"))
		if err != nil {
			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
			return
		}
		resp, md, err := local_request_LangYaPlatform_ServiceContractBookSave_0(ctx, inboundMarshaler, server, req, pathParams)
		md.HeaderMD, md.TrailerMD = metadata.Join(md.HeaderMD, stream.Header()), metadata.Join(md.TrailerMD, stream.Trailer())
		ctx = runtime.NewServerMetadataContext(ctx, md)
		if err != nil {
			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
			return
		}

		forward_LangYaPlatform_ServiceContractBookSave_0(ctx, mux, outboundMarshaler, w, req, resp, mux.GetForwardResponseOptions()...)

	})

	return nil
}

// RegisterLangYaPlatformHandlerFromEndpoint is same as RegisterLangYaPlatformHandler but
// automatically dials to "endpoint" and closes the connection when "ctx" gets done.
func RegisterLangYaPlatformHandlerFromEndpoint(ctx context.Context, mux *runtime.ServeMux, endpoint string, opts []grpc.DialOption) (err error) {
	conn, err := grpc.Dial(endpoint, opts...)
	if err != nil {
		return err
	}
	defer func() {
		if err != nil {
			if cerr := conn.Close(); cerr != nil {
				grpclog.Infof("Failed to close conn to %s: %v", endpoint, cerr)
			}
			return
		}
		go func() {
			<-ctx.Done()
			if cerr := conn.Close(); cerr != nil {
				grpclog.Infof("Failed to close conn to %s: %v", endpoint, cerr)
			}
		}()
	}()

	return RegisterLangYaPlatformHandler(ctx, mux, conn)
}

// RegisterLangYaPlatformHandler registers the http handlers for service LangYaPlatform to "mux".
// The handlers forward requests to the grpc endpoint over "conn".
func RegisterLangYaPlatformHandler(ctx context.Context, mux *runtime.ServeMux, conn *grpc.ClientConn) error {
	return RegisterLangYaPlatformHandlerClient(ctx, mux, NewLangYaPlatformClient(conn))
}

// RegisterLangYaPlatformHandlerClient registers the http handlers for service LangYaPlatform
// to "mux". The handlers forward requests to the grpc endpoint over the given implementation of "LangYaPlatformClient".
// Note: the gRPC framework executes interceptors within the gRPC handler. If the passed in "LangYaPlatformClient"
// doesn't go through the normal gRPC flow (creating a gRPC client etc.) then it will be up to the passed in
// "LangYaPlatformClient" to call the correct interceptors.
func RegisterLangYaPlatformHandlerClient(ctx context.Context, mux *runtime.ServeMux, client LangYaPlatformClient) error {

	mux.Handle("GET", pattern_LangYaPlatform_ServiceContractBook_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
		ctx, cancel := context.WithCancel(req.Context())
		defer cancel()
		inboundMarshaler, outboundMarshaler := runtime.MarshalerForRequest(mux, req)
		var err error
		ctx, err = runtime.AnnotateContext(ctx, mux, req, "/app.LangYaPlatform/ServiceContractBook", runtime.WithHTTPPathPattern("/contract/book"))
		if err != nil {
			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
			return
		}
		resp, md, err := request_LangYaPlatform_ServiceContractBook_0(ctx, inboundMarshaler, client, req, pathParams)
		ctx = runtime.NewServerMetadataContext(ctx, md)
		if err != nil {
			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
			return
		}

		forward_LangYaPlatform_ServiceContractBook_0(ctx, mux, outboundMarshaler, w, req, resp, mux.GetForwardResponseOptions()...)

	})

	mux.Handle("POST", pattern_LangYaPlatform_ServiceContractBookSave_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
		ctx, cancel := context.WithCancel(req.Context())
		defer cancel()
		inboundMarshaler, outboundMarshaler := runtime.MarshalerForRequest(mux, req)
		var err error
		ctx, err = runtime.AnnotateContext(ctx, mux, req, "/app.LangYaPlatform/ServiceContractBookSave", runtime.WithHTTPPathPattern("/contract/book"))
		if err != nil {
			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
			return
		}
		resp, md, err := request_LangYaPlatform_ServiceContractBookSave_0(ctx, inboundMarshaler, client, req, pathParams)
		ctx = runtime.NewServerMetadataContext(ctx, md)
		if err != nil {
			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
			return
		}

		forward_LangYaPlatform_ServiceContractBookSave_0(ctx, mux, outboundMarshaler, w, req, resp, mux.GetForwardResponseOptions()...)

	})

	return nil
}

var (
	pattern_LangYaPlatform_ServiceContractBook_0 = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0, 2, 1}, []string{"contract", "book"}, ""))

	pattern_LangYaPlatform_ServiceContractBookSave_0 = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0, 2, 1}, []string{"contract", "book"}, ""))
)

var (
	forward_LangYaPlatform_ServiceContractBook_0 = runtime.ForwardResponseMessage

	forward_LangYaPlatform_ServiceContractBookSave_0 = runtime.ForwardResponseMessage
)